Adjusts signing button sizes and spacing
