+++ apps/judicial-system/web/src/components/PdfButton/PdfButton.css.ts
@@ -1,7 +1,11 @@
 import { style } from '@vanilla-extract/css'
 
-export const pdfLink = style({
-  ':hover': {
-    textDecoration: 'none',
-  },
+export const pdfRow = style({
+  display: 'flex',
+  justifyContent: 'space-between',
+  flexDirection: 'row',
+  alignItems: 'center',
+  cursor: 'pointer',
+  width: '100%',
+  minHeight: '64px',
 })
