--- apps/judicial-system/web/src/components/PdfButton/PdfRow.css.ts
@@ -1,11 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-export const pdfRow = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-  flexDirection: 'row',
-  alignItems: 'center',
-  cursor: 'pointer',
-  width: '100%',
-  minHeight: '64px',
-})
